+++ inlang/source-code/website/src/services/auth/src/onSignOut.ts
@@ -2,6 +2,7 @@ import type { SetStoreFunction } from "solid-js/store"
 import type { LocalStorageSchema } from "../../../services/local-storage/index.js"
 import { publicEnv } from "@inlang/env-variables"
 import { telemetryBrowser } from "@inlang/telemetry"
+import { browserAuth } from "@lix-js/client"
 
 /**
  * This function is called when the user clicks the "Sign Out" button.
@@ -14,10 +15,7 @@ export async function onSignOut(args: { setLocalStorage: SetStoreFunction<LocalS
 	if (hasExternalGitProxy) {
 		await Promise.allSettled([
 			fetch("/services/auth/sign-out", { method: "POST" }),
-			fetch(`${publicEnv.PUBLIC_GIT_PROXY_BASE_URL}/services/auth/sign-out`, {
-				method: "POST",
-				credentials: "include",
-			}),
+			browserAuth.logout(),
 		])
 	} else {
 		await fetch("/services/auth/sign-out", { method: "POST" })