+++ inlang/source-code/website/src/services/auth/src/pages/oauth-callback.page.tsx
@@ -1,7 +1,7 @@
 import { createEffect, createResource, Match, Switch } from "solid-js"
 import MaterialSymbolsCheckCircleRounded from "~icons/material-symbols/check-circle-rounded"
 import MaterialSymbolsArrowBackRounded from "~icons/material-symbols/arrow-back-rounded"
-import { getUserInfo } from "../implementation.telefunc.js"
+import { browserAuth } from "@lix-js/client"
 import { useLocalStorage } from "#src/services/local-storage/index.js"
 
 /**
@@ -13,9 +13,7 @@ import { useLocalStorage } from "#src/services/local-storage/index.js"
  * Read more https://docs.github.com/en/developers/apps/building-oauth-apps/authorizing-oauth-apps#web-application-flow
  */
 export function Page() {
-	// ! Extremely important to wrap the get user function
-	// ! see https://github.com/brillout/telefunc/issues/56#issuecomment-1397929356
-	const [userInfo] = createResource(async () => await getUserInfo())
+	const [userInfo] = createResource(browserAuth.getUser)
 	const [localStorage] = useLocalStorage()
 
 	createEffect(() => {